added module 4:

+++ store/models.py
@@ -13,7 +13,7 @@ class Customer(models.Model):
 
 class Product(models.Model):
     name = models.CharField(max_length=200, null=True)
-    price = models.FloatField()
+    price = models.DecimalField(max_digits=7, decimal_places=2)
     digital = models.BooleanField(default=False, null=True, blank=False)
     #images
     image = models.ImageField(null=True, blank=True)
@@ -48,8 +48,8 @@ class Order(models.Model):
             if i.product.digital == False:
                 shipping = True
             return shipping
-        
-            
+
+
     @property
     def get_cart_total(self):
         orderitems = self.orderitem_set.all()
